Add method to connect links that should be connected
This commit creates the addConnectorLinks method as well as the buildStupidLinks method. Before, there were some connectivity problems with some links not being connected at a common node.The addConnectorLinks method solves this problem. The buildStupidLinks method is simply previous code that was moved to a method so that it could be used more easily.