Update the R scripts that were used for debugging
This updates the run2plots.R script which is used to create analysis plots. One plot in particular shows the modality styles (and their class definitions) from the population files throughout each iteration. This analysis is out of date though, as there is not a population file created for each iteration. In addition, the beam-utilities.R script has column headers updated to correct formatting.