Create a ModeChoiceASIM mode choice algorithm
The idea behind this algorithm is to create multiple "models". Each model has a set of utility values specific to different tour purposes. Those tour purposes come from the person's plan. The list of "models" is then passed into the ModeChoiceMultinomialLogit function, where it calculates mode choice based on the current tour purpose the individual is on.