Add tour purpose files for sf-superlight scenario
Add of these files will be needed to try and run the mode choice tour purpose algorithm. Some important files include a population file that has tour purpose as an attribute of each activity, and asim-long.csv file that has utility parameters for each tour purpose, and the config file.