Create a new tour purpose choice model
The current tour purpose mode choice model algorithm does not calculate utility values in the correct way. A default value of 0 is given everytime. (Similary, the old ModeChoiceLCCM model gave default 0 values as well). This is because the models were using the ModeChoiceMulitnomialLogit structure, instead of a choice model structure. A new algorithm for tour purpose mode choice must be constructed. Currently, this is just a copy of the ModeChoiceLCCM class. The new tour purpose model will mimic functions from the ModeChoiceLCCM.