Update all utilityOf functions and add attributes method
All the utilityOf functions are replaced with utilityOf functions that correspond to the tour purpose choice model. Much of the code is copied from ModeChoiceTourPurpose.scala. Instead of using the ModeCostTimeTransfer variable, the modechoice data is used instead. The ModeChoice data is presented as a case class here as well.