Update csv code to change cost utility values to vehicle time values
The cost coefficient is dependent on the vehicle time coefficient values. To make it easier in the BEAM code, I have just duplicated the vehicleTime values into the cost variable. Then in the BEAM code, I will take that value and multiply it by some form of each users VOT.