Convert coordinates to correct coordinate system and fix ZDI error
The taz-centroids.csv file has coordinates written in EPSG: 26912. For some reason, when getting the origin and destination coordinates though, they are not in the same coordinate system. This commit transforms the data into the correct crs. In addition, this commit fixes a small error that exists when computing the ZDI.