Create dummy hov vehicle for potential use
Agents that teleport, walk, or don't have the option to choose a car alternative need a way to choose a hov alternative. This code has the first version of the code that creates a dummy hov vehicle for those agents. It still needs to be filtered for only those with no auto options, and needs parking structure added to it.