Add hov counter to distribute teleportation more evenly
The counters are used to determine if an agent who choses hov car will have the ability to switch to teleport or not. Basically, if there are more teleporters than hov car users, all hov car users remain, whereas if there are less teleporters, there is a random chance of whether or not an hov car user switches to teleport.