Write new event that tracks trip attributes
Similar to the PersonArrivalEvent, this event is triggered anytime a person arrives at their location. Since the PersonArrivalEvent is un-editable, this event class was created. The idea is that it creates an event for every trip the person goes on (at least the trips that are completed). It tracks similar information as the experienced_plans.xml files, but does so in the form of an event.