Only allow HOV to switch to HOV_TELEPORTATION under certain circumstance
Originally, this code only let the agent switch when the currentTourPurpose value was None, to align with the fact that they only have new HOV options when the currentTourPurpose was None. Unfortunately, trigger errors are occuring, and limiting it to trip 0 and when not a RideHailTrip seems to minimize the error. it does not however, get rid of the trigger error that is occuring.