Add available trip modes from tour mode capability
In ActivitySim, trip modes are based off of tour mode values. in this commit, a new csv is created that shows which trip modes correspond to which tour modes. This commit makes the necessary changes to read in this csv and interpret it. From a tour mode value, certain trip BeamModes are provided.